API update. Reviewed by Geoff Anderson

+++ client/src/main/java/io/confluent/kafka/schemaregistry/client/rest/entities/Schema.java
@@ -28,13 +28,13 @@ public class Schema implements Comparable<Schema> {
   @Min(1)
   private Integer version;
   @Min(0)
-  private Long id;
+  private Integer id;
   @NotEmpty
   private String schema;
 
   public Schema(@JsonProperty("name") String name,
                 @JsonProperty("version") Integer version,
-                @JsonProperty("id") Long id,
+                @JsonProperty("id") Integer id,
                 @JsonProperty("schema") String schema) {
     this.name = name;
     this.version = version;
@@ -63,12 +63,12 @@ public class Schema implements Comparable<Schema> {
   }
 
   @JsonProperty("id")
-  public Long getId() {
+  public Integer getId() {
     return this.id;
   }
 
   @JsonProperty("id")
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
